Ready to add functions, all added function basic test past.

+++ Assets/Scripts/GamePlaySystem/Functionality/Exp/ExpAttributeAuthoring.cs
@@ -35,7 +35,7 @@ namespace SparFlame.GamePlaySystem.Exp
         public Tier MaxTier;
         public Tier CurTier;
         public int MaxValue;
-        public int CurValue;
+        public float CurValue;
         public Entity NextTierPrefab;
     }
